Add entities in domain/... .

+++ app/src/main/java/com/windmill/rssbutler/domain/Entity.java
@@ -3,7 +3,7 @@ package com.windmill.rssbutler.domain;
 import com.j256.ormlite.field.DatabaseField;
 
 public class Entity {
-    public static final long UNKNOWN_ID = 0;
+    public static final int UNKNOWN_ID = 0;
 
     @DatabaseField(id = true, unique = true)
     protected long id;
